+++ src/main/java/com/younggeun/delivery/admin/service/AdminService.java
@@ -1,8 +1,11 @@
 package com.younggeun.delivery.admin.service;
 
-import com.younggeun.delivery.global.entity.RoleType;
+import static com.younggeun.delivery.global.entity.RoleType.ROLE_ADMIN;
+import static com.younggeun.delivery.global.exception.type.StoreErrorCode.EXISTS_SEQUENCE_EXCEPTION;
+import static com.younggeun.delivery.global.exception.type.UserErrorCode.USER_NOT_FOUND_EXCEPTION;
+
+import com.younggeun.delivery.global.exception.RestApiException;
 import com.younggeun.delivery.global.exception.impl.PasswordMismatchException;
-import com.younggeun.delivery.global.exception.impl.UserNotFoundException;
 import com.younggeun.delivery.global.model.Auth;
 import com.younggeun.delivery.store.domain.CategoryRepository;
 import com.younggeun.delivery.store.domain.dto.CategoryDto;
@@ -30,9 +33,9 @@ public class AdminService implements UserDetailsService {
   @Override
   public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
     User user = userRepository.findByEmail(email)
-        .orElseThrow(() -> new UsernameNotFoundException("User not found with email: " + email));
+        .orElseThrow(() -> new RestApiException(USER_NOT_FOUND_EXCEPTION));
 
-    if(!user.getRole().equals(RoleType.ROLE_ADMIN)) {
+    if(!user.getRole().equals(ROLE_ADMIN)) {
       throw new RuntimeException();
     }
 
@@ -45,13 +48,13 @@ public class AdminService implements UserDetailsService {
 
   // 로그인
   public User authenticate(Auth.SignIn user) {
-    var member = userRepository.findByEmail(user.getEmail()).orElseThrow(UserNotFoundException::new);
+    var member = userRepository.findByEmail(user.getEmail()).orElseThrow(() -> new RestApiException(USER_NOT_FOUND_EXCEPTION));
 
     if(!passwordEncoder.matches(user.getPassword(), member.getPassword())) {
       throw new PasswordMismatchException();
     }
 
-    if(!member.getRole().equals(RoleType.ROLE_ADMIN)) {
+    if(member.getRole() == ROLE_ADMIN) {
       throw new RuntimeException();
     }
 
@@ -59,25 +62,26 @@ public class AdminService implements UserDetailsService {
   }
 
   public List<Category> getCategory() {
-    return categoryRepository.findAll();
+    return categoryRepository.findAllByOrderBySequenceAsc();
   }
 
   public Category createCategory(CategoryDto categoryDto) {
+    if(existIdx(categoryDto.getSequence())) throw new RestApiException(EXISTS_SEQUENCE_EXCEPTION);
     return categoryRepository.save(categoryDto.toEntity());
   }
 
   public Category updateCategory(CategoryDto categoryDto, Long categoryId) {
-    return categoryRepository.save(Category.builder()
-                                            .categoryId(categoryId)
-                                            .name(categoryDto.getName())
-                                            .sequence(categoryDto.getSequence())
-                                            .build());
+    if(existIdx(categoryDto.getSequence())) throw new RestApiException(EXISTS_SEQUENCE_EXCEPTION);
 
+    return categoryRepository.save(categoryDto.toEntity(categoryId));
+  }
+
+  public boolean existIdx(int idx) {
+    return categoryRepository.existsBySequence(idx);
   }
 
   @Transactional
   public boolean deleteCategory(long categoryId) {
-
     categoryRepository.deleteById(categoryId);
     return true;
   }
