+++ src/main/java/com/younggeun/delivery/global/exception/type/StoreErrorCode.java
@@ -11,6 +11,8 @@ public enum StoreErrorCode implements ErrorCode {
   STORE_NOT_FOUND(HttpStatus.BAD_REQUEST, "해당 가게를 찾을 수 없습니다."),
   STORE_CATEGORY_NOT_FOUND(HttpStatus.BAD_REQUEST, "해당 카테고리를 찾을 수 없습니다."),
   MENU_NOT_FOUND(HttpStatus.BAD_REQUEST, "해당 메뉴를 찾을 수 없습니다."),
+  EXISTS_SEQUENCE_EXCEPTION(HttpStatus.BAD_REQUEST, "이미 존재하는 순번입니다."),
+  MISMATCH_PARTNER_STORE(HttpStatus.BAD_REQUEST, "가게에 등록된 파트너와 로그인한 파트너가 일치하지 않습니다.."),
   MISMATCH_STORE_CATEGORY(HttpStatus.BAD_REQUEST, "해당 가게와 카테고리의 가게가 일치하지 않습니다."),
   ALREADY_EXIST_SEQUENCE(HttpStatus.BAD_REQUEST, "해당 순서로 등록할 수 없습니다.");
   private final HttpStatus httpStatus;
