+++ src/main/java/com/younggeun/delivery/store/domain/CategoryRepository.java
@@ -1,8 +1,10 @@
 package com.younggeun.delivery.store.domain;
 
 import com.younggeun.delivery.store.domain.entity.Category;
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
-
+  List<Category> findAllByOrderBySequenceAsc();
+  boolean existsBySequence(int idx);
 }
