+++ src/main/java/com/younggeun/delivery/store/service/StoreService.java
@@ -1,6 +1,10 @@
 package com.younggeun.delivery.store.service;
 
-import com.younggeun.delivery.global.exception.impl.UserNotFoundException;
+import static com.younggeun.delivery.global.exception.type.CommonErrorCode.FILE_SAVE_ERROR;
+import static com.younggeun.delivery.global.exception.type.StoreErrorCode.MISMATCH_PARTNER_STORE;
+import static com.younggeun.delivery.global.exception.type.UserErrorCode.USER_NOT_FOUND_EXCEPTION;
+
+import com.younggeun.delivery.global.exception.RestApiException;
 import com.younggeun.delivery.partner.domain.PartnerRepository;
 import com.younggeun.delivery.partner.domain.entity.Partner;
 import com.younggeun.delivery.store.domain.CategoryRepository;
@@ -13,7 +17,12 @@ import com.younggeun.delivery.store.domain.entity.Category;
 import com.younggeun.delivery.store.domain.entity.Store;
 import com.younggeun.delivery.store.domain.entity.StorePhoto;
 import com.younggeun.delivery.store.domain.entity.StoreProfilePhoto;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.time.LocalDate;
 import java.util.Objects;
+import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -21,6 +30,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.FileCopyUtils;
+import org.springframework.web.multipart.MultipartFile;
 
 @Slf4j
 @Service
@@ -33,7 +44,7 @@ public class StoreService {
   private final CategoryRepository categoryRepository;
 
   public Page<StoreDto> selectPartnerStore(Authentication authentication, Pageable pageable) {
-    Partner partner = partnerRepository.findByEmail(authentication.getName()).orElseThrow(UserNotFoundException::new);
+    Partner partner = getPartner(authentication);
     Page<Store> stores = storeRepository.findAllByPartner(partner, pageable);
 
     return stores.map(store -> {
@@ -54,7 +65,7 @@ public class StoreService {
 
   @Transactional
   public Store createStore(Authentication authentication, StoreDto request) {
-    Partner partner = partnerRepository.findByEmail(authentication.getName()).orElseThrow(UserNotFoundException::new);
+    Partner partner = getPartner(authentication);
     Category category = categoryRepository.findById(request.getCategoryId()).orElseThrow(RuntimeException::new);
 
     return storeRepository.save(Store.builder()
@@ -78,29 +89,104 @@ public class StoreService {
         .partner(partner).build());
   }
 
-  public StorePhoto createStorePhoto(Authentication authentication, PhotoDto storePhotoDto, Long storeId) {
-    Partner partner = partnerRepository.findByEmail(authentication.getName()).orElseThrow(UserNotFoundException::new);
+  public StorePhoto createStorePhoto(Authentication authentication, MultipartFile file, Long storeId,
+      String photoBaseLocalPath, String photoBaseUrlPath) {
+    Partner partner = getPartner(authentication);
     Store store = storeRepository.findById(storeId).orElseThrow();
 
     if(!Objects.equals(store.getPartner().getPartnerId(), partner.getPartnerId())) {
       throw new RuntimeException();
     }
+
+    PhotoDto storePhotoDto = savePhotos(file, photoBaseLocalPath, photoBaseUrlPath);
 
     storePhotoDto.setStore(store);
 
     return storePhotoRepository.save(storePhotoDto.toStorePhotoEntity());
   }
 
-  public StoreProfilePhoto createStoreProfilePhoto(Authentication authentication, PhotoDto storeProfilePhotoDto, long storeId) {
-    Partner partner = partnerRepository.findByEmail(authentication.getName()).orElseThrow(UserNotFoundException::new);
+  public StoreProfilePhoto createStoreProfilePhoto(Authentication authentication, MultipartFile file, long storeId,
+      String profileBaseLocalPath, String profileBaseUrlPath) {
+    Partner partner = getPartner(authentication);
     Store store = storeRepository.findById(storeId).orElseThrow();
 
     if(!Objects.equals(store.getPartner().getPartnerId(), partner.getPartnerId())) {
-      throw new RuntimeException();
+      throw new RestApiException(MISMATCH_PARTNER_STORE);
     }
+
+    PhotoDto storeProfilePhotoDto = savePhotos(file, profileBaseLocalPath, profileBaseUrlPath);
 
     storeProfilePhotoDto.setStore(store);
 
     return storeProfilePhotoRepository.save(storeProfilePhotoDto.toStoreProfilePhotoEntity());
+  }
+
+  private PhotoDto savePhotos(MultipartFile file, String localPath, String urlPath) {
+    PhotoDto storePhotoDto = new PhotoDto();
+    String saveFilename = "";
+    String urlFilename = "";
+
+    if (file != null) {
+      String originalFilename = file.getOriginalFilename();
+
+      String[] arrFilename = getNewSaveFile(localPath, urlPath, originalFilename);
+
+      saveFilename = arrFilename[0];
+      urlFilename = arrFilename[1];
+
+      try {
+        File newFile = new File(saveFilename);
+        FileCopyUtils.copy(file.getInputStream(), new FileOutputStream(newFile));
+      } catch (IOException e) {
+        log.warn(e.getMessage());
+        throw new RestApiException(FILE_SAVE_ERROR);
+      }
+    }
+
+    storePhotoDto.setUrl(urlFilename);
+    storePhotoDto.setPhotoName(saveFilename);
+
+    return storePhotoDto;
+  }
+
+  private Partner getPartner(Authentication authentication) {
+    return partnerRepository.findByEmail(authentication.getName()).orElseThrow(() -> new RestApiException(USER_NOT_FOUND_EXCEPTION));
+  }
+
+  private String[] getNewSaveFile(String baseLocalPath, String baseUrlPath, String originalFilename) {
+
+    LocalDate now = LocalDate.now();
+
+    String[] dirs = {
+        String.format("%s/%d/", baseLocalPath,now.getYear()),
+        String.format("%s/%d/%02d/", baseLocalPath, now.getYear(),now.getMonthValue()),
+        String.format("%s/%d/%02d/%02d/", baseLocalPath, now.getYear(), now.getMonthValue(), now.getDayOfMonth())};
+
+    String urlDir = String.format("%s/%d/%02d/%02d/", baseUrlPath, now.getYear(), now.getMonthValue(), now.getDayOfMonth());
+
+    for(String dir : dirs) {
+      File file = new File(dir);
+      if (!file.isDirectory()) {
+        file.mkdir();
+      }
+    }
+
+    String fileExtension = "";
+    if (originalFilename != null) {
+      int dotPos = originalFilename.lastIndexOf(".");
+      if (dotPos > -1) {
+        fileExtension = originalFilename.substring(dotPos + 1);
+      }
+    }
+
+    String uuid = UUID.randomUUID().toString().replaceAll("-", "");
+    String newFilename = String.format("%s%s", dirs[2], uuid);
+    String newUrlFilename = String.format("%s%s", urlDir, uuid);
+    if (fileExtension.length() > 0) {
+      newFilename += "." + fileExtension;
+      newUrlFilename += "." + fileExtension;
+    }
+
+    return new String[]{newFilename, newUrlFilename};
   }
 }
